% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXvcf.R
\name{addAnno}
\alias{addAnno}
\title{Read annotation file}
\usage{
addAnno(dt, anno_file)
}
\arguments{
\item{dt}{A \code{data.table} object.}

\item{anno_file}{A \code{character}. The name of a file containing annotations.}
}
\value{
A \code{data.table} object that contains allelic coverage, genotype
and annotations at the covered SNPs.
}
\description{
Read a given annotation file and merge it with a data.table containing the
relevant information to exstimate inactivated X chromosome expression.
}

