% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotates_snps}
\alias{annotates_snps}
\title{Annotate}
\usage{
annotates_snps(object, pos_col = "POS", chr = "X")
}
\arguments{
\item{object}{A \code{data.table} with genomic positions. Typically, the 
table returned by \code{betaBinomXI}.}

\item{pos_col}{A \code{character} indicating the column that contains genomic
positions.}

\item{chr}{A \code{character}. Filter SNPs for the selected chromosome.}
}
\description{
Find SNPs based on matching position in dbgap using biomaRt
}
