% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabin.R
\name{betaBinomXI}
\alias{betaBinomXI}
\title{Fit mixture model}
\usage{
betaBinomXI(genic_dt, method = "all", type = "midp", plot = FALSE,
  hist = FALSE, graph_summary = FALSE, xciGenes = NULL, limits = F)
}
\description{
Fit a mixture model to estimate mosaicism and XCI-escape.
}
