% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{betaParam}
\alias{betaParam}
\title{Converting beta distribution parameters}
\usage{
betaParam(alpha = NULL, beta = NULL, m = NULL, theta = NULL,
  mu = NULL, sigma2 = NULL)
}
\arguments{
\item{alpha}{First shape parameter}

\item{beta}{Second shape parameter}

\item{m}{Mode}

\item{theta}{Concentration}

\item{sigma2}{Variance}

\item{mean}{Mean}
}
\value{
A \code{list} with the two new parameters.
}
\description{
Convert parameter values between different beta distribution parametrization
}
\examples{

betaParam(alpha = 5, beta = 5)
betaParam(m = 0.5, theta = 10)
betaParam(mu = 0.5, sigma2 = 0.02272727)

}
