% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_escape.R
\name{CMH}
\alias{CMH}
\title{Cochran-Mantel-Haenszel test}
\usage{
CMH(calls, condition, value)
}
\arguments{
\item{calls}{A \code{character}. XCI status prediction as outputed by
\code{betaBinomXI}.}

\item{condition}{A \code{character}. The column name that contains the
outcomes of interest.}

\item{value}{A \code{character}. The outcome of interest. All other values
in the \code{condition} columns will be matched against it.}
}
\description{
Runs CMH on \code{XCIR} calls to test the association between XCI status
and an outcome of interest.
}
\note{
The CMH will only be computed using genes that have at least 1 sample for
each outcome.
}
\seealso{
\code{\link{betaBinomXI}}
}
