% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{consensusXCI}
\alias{consensusXCI}
\title{XCI consensus}
\usage{
consensusXCI(redownload = F, simple = T)
}
\arguments{
\item{redownload}{A \code{logical}. If set to TRUE, the original supplementary
file is redownloaded from PMC.}

\item{simple}{A \code{logical}. If set to TRUE, minimal information is returned,
only for genes with an available XCIR classification.}
}
\description{
Read consensus & XCIR calls for all X-linked genes
}
\details{
The consensus is as published in Supplementary table S1 of
Balaton et al. (Biol Sex Differ. 2015). doi: 10.1186/s13293-015-0053-7
}
