% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underdev.R
\name{fisherCombine}
\alias{fisherCombine}
\title{Combine p-values from multiple SNPs}
\usage{
fisherCombine(newcalls, maxSNP = 4, method = "fisher")
}
\arguments{
\item{newcalls}{The \code{data.table} output by \code{newCalls}.}

\item{maxSNP}{A \code{numeric}. The maximum number of SNPs to use (i.e:
p-values to combine).}

\item{method}{A \code{character}. One of "fisher", "stouffer", "zscore".
zscore uses the total read count to weight each SNP.}
}
\value{
The \code{newcalls} \code{data.table} with a new f_value column.
}
\description{
Combine p-values from multiple SNPs
}
