% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCI.R
\name{getCellFrac}
\alias{getCellFrac}
\title{Get fraction of cell expressing the inactivated allele}
\usage{
getCellFrac(dt, xciGenes)
}
\arguments{
\item{dt}{A \code{data.table}. The data.}

\item{xciGenes}{A \code{character}. The name of the genes to analyze.}
}
\value{
A \code{data.table} containing the average, and median fraction of
expression of the inactivated chromosome for each subject.
}
\description{
Get fraction of cell expressing the inactivated allele
}

