% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellFrac.R
\name{getCellFrac2}
\alias{getCellFrac2}
\title{Get Cell fraction}
\usage{
getCellFrac2(dt, xciGenes = NULL, method = "naive", xist_only = TRUE,
  rm_inac = 2, plot = FALSE)
}
\arguments{
\item{dt}{A \code{data.table}. The output of \code{getGenicDP}.}
}
\description{
No longer used
}

