% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getError.R
\name{getError}
\alias{getError}
\title{Plot error and power}
\usage{
getError(xi, alpha = 0.05, escape_cutoff = 1, inac_cutoff = 2,
  plot = TRUE, xciGenes = TRUE)
}
\arguments{
\item{xi}{A \code{data.table}. Escape calls, as outputted by \code{betaBinomXI}.}

\item{alpha}{A \code{numeric}. The significance level.}

\item{escape_cutoff}{A \code{numeric}. A gene must have more reads than the
cutoff on both alleles in the two skewed samples to be considered as
escaping XCI.}

\item{inac_cutoff}{A \code{numeric}. A gene must have at most that many reads
on both alleles in the two skewed samples to be considered as subject to
inactivation.}

\item{plot}{A \code{logical}. If set to false, return Type1 error and power
table without plotting barplots.}

\item{xciGenes}{A \code{logical}. If set to TRUE, uses genes from the
training set as inactivated genes (still subject to filtering via
\code{inac_cutoff}). If set to FALSE, a hand curated gene list will be used
insteda.}
}
\description{
Plot power and type 1 error for each sample. Using the most skewed samples
as a reference. Used to asses the quality of the prediction.
}
\note{
This will only work if all samples belong to the same subject and two of the
samples are mono-allelicly expressed.
}
