% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenic.R
\name{getGenicDP}
\alias{getGenicDP}
\title{Get expression at the gene level}
\usage{
getGenicDP(dt_anno, highest_expr = FALSE, pool = FALSE,
  gender_file = NULL)
}
\arguments{
\item{dt_anno}{A \code{data.table}. An annotated table of read counts for
each SNP, as outputted by \code{addAnno}}

\item{pool}{A \code{logical}. Only works when \code{highest_expr} is set to
TRUE. If set to TRUE, the read counts are pooled accross all samples for
each SNP. Only use this if the samples come from the same subject}
}
\value{
A \code{data.table}. That should be used as input for
 \code{betaBinomXI}.
}
\description{
Calculate allele specific expression for each gene in each sample, either
using only the most expressed SNP or using all SNPs (when phasing has been
performed).
}
\seealso{
betaBinomXI, addAnno
}
