% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getXCIstate}
\alias{getXCIstate}
\title{Classify X-genes}
\usage{
getXCIstate(xciObj)
}
\arguments{
\item{xciObj}{A \code{data.table}. The table returned by \code{betaBinomXI}}
}
\value{
A \code{data.table} with genes and their XCI-state.
}
\description{
Classify X-linked genes between Escape (E), Variable Escape (VE) and Silenced (S)
}
