% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCI.R
\name{getXIexpr}
\alias{getXIexpr}
\title{Estimate inactivated X chromosome expression}
\usage{
getXIexpr(dt, dt_frac, dp_max = NULL, fix_phasing = TRUE, output = FALSE,
  cutoff = NA)
}
\arguments{
\item{dt}{A \code{data.table}. The data. It must contain a gender column.}

\item{dt_frac}{A \code{data.table}. The result of the \code{getCellFrac}
function.}

\item{dp_max}{A \code{numeric}. If any gene has a total coverage above this
number, the coverage will be set to this number.}

\item{output}{A \code{logical}. If set to TRUE, output a table for each
cutoff level.}

\item{cutoffs}{A \code{numeric} or NA.}
}
\description{
Calculate inactivated X chromosome expression.
}
\details{
Select the females. calculate the proportion of expression coming from the
inactivated X chromosome for each gene.

When the allelic coverage is very big for some genes, it gives more power,
which allow to spot smaller significant differences. When only major
differences are of interest, the \code{dp_max} parameter can be set to a
lower value.
}

