% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{perm_test}
\alias{perm_test}
\title{Perform a permutation test}
\usage{
perm_test(table, y, nperm = 100, tau_cutoff = NULL, varweight = T,
  verbose = F)
}
\arguments{
\item{table}{A \code{data.table}. The output of \code{betaBinomXI}.}

\item{y}{A \code{character}. A column of \code{table}.}

\item{nperm}{A \code{numeric}. The number of permutations to perform.}

\item{tau_cutoff}{A \code{numeric}. Optional argument, if set to a numeric
value, all taus less than the cutoff will be set tot 0.}

\item{varweight}{A \code{logical}. If set to TRUE, taus are weighted by the
inverse of their variance.}

\item{verbose}{A \code{logical}. If set to TRUE, will post additional information
about the permutations.}
}
\value{
A \code{data.table} with the following information
\item{GENE}{The gene symbol}
\item{nsamples}{The number of samples used for the test}
\item{Tobs}{The test statistic}
\item{perm_pvalue}{The p-value}
}
\description{
Permutation test to assess the effect of the estimated Xi expression
(represented as tau, the fration of the total expression) on a phenotype.
}
