% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_escape.R
\name{plot_status}
\alias{plot_status}
\title{Plot escape across X}
\usage{
plot_status(sub_xi, alpha = 0.05, min_sup = 0, rownames = NULL,
  inference = "asymptotic", nsamples = F, cutoffs = c(25, 75),
  theme = "gray", hg = 19)
}
\arguments{
\item{sub_xi}{A \code{data.table}. The output of \code{betaBinomXI}.}

\item{alpha}{A \code{numeric} between 0 and 1. The significance level for
escape calls.}

\item{min_sup}{A \code{numeric}. Only genes with at least that many samples
will be displayed.}

\item{rownames}{A \code{logical} or NULL. If set to NULL, gene names will
be displayed only if there are less than 100 genes on the plot.}

\item{inference}{A \code{character}. Which of the \code{betaBinomXI}
predictions should be used, "asymptotic" or "exact".}

\item{nsamples}{A \code{logical}. If TRUE, display the total number of samples
in the dataset.}

\item{cutoffs}{A \code{numeric} vector of length 2. The boundaries for
calling subject, variable and escape genes.}

\item{theme}{A \code{character}. Available options are "gray", "status" and
"samplecount".}

\item{hg}{A \code{numeric}. The genome version that was used to align the
RNA-Seq data. This is used to display annotations.}
}
\value{
An invisible \code{data.table} with the following columns
\item{GENE}{Gene symbol}
\item{N_support}{The number of samples for which a prediction was made for that gene}
\item{N_escape}{The number of samples where this gene was found to escape}
\item{pe}{The fraction of escape, N_escape/N_support}
}
\description{
Barplot of the escape status of X-linked genes. Each bar represent the
proportions of samples in which a gene is found to escape.
Genes are ordered along the X chromosome.
}
