% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_escape.R
\name{plot_status_fraction}
\alias{plot_status_fraction}
\title{Plot the fraction of genes in each category}
\usage{
plot_status_fraction(xi, alpha = 0.05, inference = "asymptotic",
  threshold = c(0.25, 0.75), min_support = 0, plot = TRUE)
}
\arguments{
\item{xi}{A \code{data.table}. The calls, output from \code{betaBinomXI}.}

\item{alpha}{A \code{numeric}. The significance level.}

\item{inference}{A \code{character}. Which p-value should be used to
categorize each gene. Valid choices are 'asymptotic' and 'exact'.}

\item{threshold}{A \code{character} of length 2. Genes with a percent of
escape lower than the first number will be considered subject to
inactivation. Genes with a higher percent of escape than the second number
will be categorized as escaping inactivation.}

\item{min_support}{A \code{numeric}. Only calculate percentage of escape for
genes that have at least that many samples for which a prediction was made.}
}
\value{
An invisible \code{data.table} with one row for each gene.
}
\description{
Barplot of the fraction of gene that are 'subject', 'variable' and 'escape'
inactivation from the calls made in XCIR
}
