% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabin.R
\name{plotBBCellFrac}
\alias{plotBBCellFrac}
\title{Plot cell fraction estimates}
\usage{
plotBBCellFrac(xci_dt, xcig = T, gene_names = "")
}
\description{
Plot cell fraction estimates from list of known XCI genes
}
\details{
This function is mostly used in \code{betaBinomXI} to ensure that the cell
fraction is estimated properly. However, it can be used from the output
of \code{betaBinomXI} to troubleshoot estimation issues.
}
