% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXvcf.R
\name{readXVcf}
\alias{readXVcf}
\title{Read a vcf file to extract information relevant to XCI}
\usage{
readXVcf(vcf_file)
}
\arguments{
\item{vcf_file}{A \code{character}. The path to a vcf file.}
}
\value{
A \code{data.table} object in long format of length sample*site.
}
\description{
Read a given vcf file and extract minimum information required for the
estimation of the expression of inactivated X chromosome.
}

