% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_escape.R
\name{summescape}
\alias{summescape}
\title{Summarize escape frequency}
\usage{
summescape(table, fmax = 0.5, alpha = 0.05)
}
\arguments{
\item{table}{A \code{data.table}. The output of \code{betaBinomXI}.}

\item{fmax}{A \code{numeric}. The maximum skew (takes values between 0 and
0.5, 0.5 corresponds to balanced samples).}

\item{alpha}{A \code{numeric}. The significance level of escape calls.}
}
\description{
Summarize escape frequency
}
