% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabin.R
\name{xci_calls}
\alias{xci_calls}
\title{Call XCI-states}
\usage{
xci_calls(xcirtable)
}
\arguments{
\item{xcirtable}{A \code{data.table}. The results of \code{betaBinomXI}. 
Where the skewing columns \code{f} has been manually modified.}
}
\value{
A \code{data.table} with updated XCI-inference for all genes/samples.
}
\description{
Use precomputed skewing and ASE ratios to test for escape
}
\details{
This can be used to recompute XCI-states using a pre-determined skewing
(e.g: experimentally set, or measured with RNA-qsnapshot).
}
\seealso{
\code{betaBinomXI}
}
