% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCI.R
\name{XCIanalyze}
\alias{XCIanalyze}
\title{Run the whole pipeline in one function}
\usage{
XCIanalyze(vcf_file, haps_file, fix_phasing = TRUE, add_gender = TRUE)
}
\description{
Run the whole pipeline in one function
}

