% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCIR.R
\name{XCIR-package}
\alias{XCIR-package}
\alias{XCIR}
\title{Estimating inactivated X chromosome expression}
\description{
Tools for the analysis of X chromosome inactivation (XCI) and XCI-escape inference.
}
\author{
Renan Sauteraud <rxs575@psu.edu>
}
