% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXvcf.R
\name{addAnno}
\alias{addAnno}
\title{Read annotation file}
\usage{
addAnno(dt, seqm_annotate = TRUE, read_count_cutoff = 20,
  filter_mono_cutoff = 3, anno_file = NULL)
}
\arguments{
\item{dt}{A \code{data.table} object.}

\item{seqm_annotate}{A \code{logical}. If set to TRUE, the \code{seqminer}
package will be used to annotate \code{dt}. If set to FALSE, this function
is a simple read count filtering step.}

\item{read_count_cutoff}{A \code{numeric}. Keep only SNPs that have at least
that many reads.}

\item{filter_mono_cutoff}{A \code{numeric}. If > 0, remove mono-allelic
SNPs. See details for more information.}

\item{anno_file}{A \code{character}. The name of a file containing annotations.}
}
\value{
A \code{data.table} object that contains allelic coverage, genotype
and annotations at the covered SNPs.
}
\description{
Read a given annotation file and merge it with a data.table containing the
relevant information to exstimate inactivated X chromosome expression and
filter out SNPs with low coverage.
}
\details{
\code{filter_mono_cutoff} will remove SNPs that do not have at least one read
covering each allele accross the pooled samples. i.e: If any sample has at
least one read covering the first allele and any sample has at least one
read covering the second allele, then the SNP will be preserved.
}
