% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabin.R
\name{betaBinomXI}
\alias{betaBinomXI}
\title{Fit mixture model}
\usage{
betaBinomXI(genic_dt, method = "all", plot = FALSE, hist = FALSE,
  graph_summary = FALSE, xciGenes = NULL, limits = FALSE)
}
\arguments{
\item{genic_dt}{A \code{data.table}. The table as outputted by \code{getGenicDP}.}

\item{method}{A \code{character} indicating which model to use to estimate
the mosaicism. Valid choices are "all", "BB", "MM", "MM2", "MM3". See details.}

\item{plot}{A \code{logical}. If set to TRUE, information about the training
set and the skewing estimate will be plotted.}

\item{hist}{A \code{logical}. If set to TRUE, an histogram of the skewing
estimates will be displayed.}

\item{graph_summary}{A \code{logical}. Unused. Will include better summary of
the fit.}

\item{xciGenes}{A \code{character} or NULL. To be passed to \code{readXCI} to
select the training set of inactivated genes.}

\item{limits}{A \code{logical}. If set to TRUE, the optimization will be
constrained. Using upper bounds on the probability of sequencing error and
escape in the training set ensures that the dominant mixture represents the
skewing for inactivated genes.}
}
\description{
Fit a mixture model to estimate mosaicism and XCI-escape.
}
\details{
The method determines the number of components used in the mixture model. By
default, "all" tries all combinations of mixtures and the best estimate is
kept using backward selection based on AIC.
BB is a simple beta-binomial. MM adds a binomial component to model the
sequencing errors. MM2 jointly models the probability of misclasification
in the training set. MM3 include all 3 components.
}
\seealso{
getGenicDP readXCI
}
