% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabin.R
\name{betaBinomXI}
\alias{betaBinomXI}
\title{Fit mixture model}
\usage{
betaBinomXI(genic_dt, model = "AUTO", plot = FALSE, hist = FALSE,
  flag = 0, xciGenes = NULL, limits = TRUE)
}
\arguments{
\item{genic_dt}{A \code{data.table}. The table as outputted by \code{getGenicDP}.}

\item{model}{A \code{character} indicating which model to use to estimate
the mosaicism. Valid choices are "AUTO", "BB", "MM", "MM2", "MM3". See details.}

\item{plot}{A \code{logical}. If set to TRUE, information about the training
set and the skewing estimate will be plotted.}

\item{hist}{A \code{logical}. If set to TRUE, an histogram of the skewing
estimates will be displayed.}

\item{flag}{A \code{numeric}. Specify how to handle convergence issues. See
details.}

\item{xciGenes}{A \code{character} or NULL. To be passed to \code{readXCI} to
select the training set of inactivated genes.}

\item{limits}{A \code{logical}. If set to TRUE, the optimization will be
constrained. Using upper bounds on the probability of sequencing error and
escape in the training set ensures that the dominant mixture represents the
skewing for inactivated genes.}
}
\description{
Fit a mixture model to estimate mosaicism and XCI-escape.
}
\details{
The model determines the number of components used in the mixture model. By
default, "AUTO" tries all combinations of mixtures and the best estimate is
kept using backward selection based on AIC.
BB is a simple beta-binomial. MM adds a binomial component to model the
sequencing errors. MM2 jointly models the probability of misclasification
in the training set. MM3 include all 3 components.

Flags in the output reports issues in convergence. If \code{flag} is set to 0,
nothing is done. If set to 1, the model selection will avoid flagged models
(will favor parcimonious models).
If set to 2, calls for which the best selected model had convergence issue
will be removed.
}
\seealso{
getGenicDP readXCI
}
