% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{betaAB}
\alias{betaAB}
\alias{betaMC}
\alias{betaMV}
\title{Converting beta distribution parameters}
\usage{
betaAB(mean, var)

betaMC(alpha, beta)

betaMV(alpha, beta)
}
\arguments{
\item{mean}{Mean}

\item{var}{Variance}

\item{alpha}{First shape parameter}

\item{beta}{Second shape parameter}

\item{m}{Mode}

\item{theta}{Concentration}
}
\value{
A \code{list} with the two new parameters.
}
\description{
These functions convert parameter values between different beta distribution parametrization
}
\examples{

betaMV(1, 1)
betaMV(1, 10)
betaMV(10, 10)
# In a data.table (such as XCIR output)
dt <- data.table(it = LETTERS[1:3], a_est = c(1, 1, 10), b_est = c(1, 10, 10))
dt[, `:=`(c("mu", "s2"), betaMV(a_est, b_est))]


}
