% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCI.R
\name{getCellFrac}
\alias{getCellFrac}
\title{Get fraction of cell expressing the inactivated allele}
\usage{
getCellFrac(dt, readCountThresh = 10, fix_phasing = TRUE,
  overall_mean = F, xist_only = F, merge_xist = TRUE, xciGenes = NULL,
  plot = FALSE)
}
\arguments{
\item{dt}{A \code{data.table}. The output of \code{getGenicDP}.}

\item{fix_phasing}{A \code{logical}. If set to TRUE, the fraction above .5
are set to 1-frac.}

\item{xciGenes}{A \code{character}. A list or a filename that contains a
list of known XCI genes.}

\item{plot}{A \code{logical}. If set to TRUE, a plot of the median fraction
for each sample will be generated.}
}
\value{
A \code{data.table} containing the average, and median fraction of
expression of the inactivated chromosome for each sample. frac_XIST is the
fraction of total XIST expression contributed by the inactive X chromosome.
}
\description{
Get fraction of cell expressing the inactivated allele
}
\details{
The default xciGenes is a list of 177 genes known to be affected by X
chromosome inactivation from Cotton et al. 2013
(dx.doi.org/10.1186/gb-2013-14-11-r122).
}
\seealso{
getGenicDP
}
