% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getError.R
\name{getError}
\alias{getError}
\title{Plot error and power}
\usage{
getError(xi, alpha = 0.05, escape_cutoff = 1, inac_cutoff = 2,
  plot = TRUE, xciGenes = TRUE)
}
\description{
Plot power and type 1 error for each sample. Using the most skewed samples
as a reference. Used to asses the quality of the prediction.
}
\note{
This will only work if all samples belong to the same subject.
}
