% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getErrors.R
\name{getError}
\alias{getError}
\title{Get proportion of errors}
\usage{
getError(xi, frac_type = NULL, alpha = 0.05, cutoff = 1, plot = TRUE,
  xciGenes = NULL)
}
\arguments{
\item{xi}{A \code{data.table}. The result of \code{getXIexpr}. The predicted
status for each gene and sample.}

\item{frac_type}{A \code{character}, can be NULL, "mean", "median" or "XIST"
NULL defaults to "mean".}

\item{alpha}{A \code{numeric}. The significance threshold.}

\item{cutoff}{A \code{numeric}. Escaped genes are genes for which the
inactivated copy has more reads than this cutoff.}

\item{plot}{A \code{logical}. Create a barplot of the Type 1 error and power.}

\item{min_read}{A \code{numeric}. The minimum number of reads for a gene to
be considered informative.}
}
\description{
Get the proportion of type 1 and type 2 error for each sample for selected
cell fraction type
}

