% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenicDP.R
\name{getGenicDP}
\alias{getGenicDP}
\title{Get sum of the coverage of all SNPs for each gene}
\usage{
getGenicDP(dt_anno, highest_expr = FALSE, pool = FALSE,
  gender_file = NULL)
}
\arguments{
\item{pool}{A \code{logical}. Only works when \code{highest_expr} is set to
TRUE. If set to TRUE, the read counts are pooled accross all samples for
each SNP. Only use this if the samples come from the same subject}
}
\description{
Get sum of the coverage of all SNPs for each gene
}
