% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCI.R
\name{getGenicDP}
\alias{getGenicDP}
\title{Get sum of the coverage of all SNPs for each gene}
\usage{
getGenicDP(dt_anno, highest_expr = FALSE, plot = FALSE,
  gender_file = NULL)
}
\arguments{
\item{plot}{A \code{logical}. If set to TRUE, a histogram of the gene
coverage will be displayed.}

\item{gender_file}{A \code{character}. An optional file that contains
matching for sample to gender. Must have two columns.}
}
\description{
Get sum of the coverage of all SNPs for each gene
}

