% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCI.R
\name{getGenicDP}
\alias{getGenicDP}
\title{Get sum of the coverage of all SNPs for each gene}
\usage{
getGenicDP(dt_anno, plot = FALSE, gender_file = NULL)
}
\description{
Get sum of the coverage of all SNPs for each gene
}

