% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getXCIstate}
\alias{getXCIstate}
\title{Classify X-genes}
\usage{
getXCIstate(xciObj)
}
\arguments{
\item{xciObj}{A \code{data.table}. The table returned by \code{betaBinomXI}}
}
\value{
A \code{data.table} with genes and their XCI-state.
}
\description{
Classify X-linked genes between Escape (E), Variable Escape (VE) and Silenced (S)
}
\examples{
library(data.table)
# Simulated data
dtf <- system.file("extdata/data2_vignette.tsv", package = "XCIR")
dt <- fread(dtf)
xcigf <- system.file("extdata/xcig_vignette.txt", package = "XCIR")
xcig <- readLines(xcigf)
# Run all models on the data
all <- betaBinomXI(dt, xciGenes = xcig)
# Simple BetaBinomial model and show histogram of skewing
bb <- betaBinomXI(dt, xciGenes = xcig, model = "BB", hist = TRUE)

# Plotting fits
stoshow <- paste0("sample", c(31, 33, 35, 40)) #interesting samples
plotQC(all[sample \%in\% stoshow], xcig = xcig)

# Summarizing results
# Sample information
samps <- sample_clean(all)
# Gene-level predictions
xcistates <- getXCIstate(all)
}
