% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_escape.R
\name{plot_status_fraction}
\alias{plot_status_fraction}
\title{Plot the fraction of genes in each category}
\usage{
plot_status_fraction(xi, alpha = 0.05, inference = "asymptotic",
  threshold = c(0.25, 0.75), min_support = 0, plot = TRUE)
}
\arguments{
\item{xi}{A \code{data.table}. The calls, output from \code{betaBinomXI}.}

\item{alpha}{A \code{numeric}. The significance level.}

\item{inference}{A \code{character}. Which p-value should be used to
categorize each gene. Valid choices are 'asymptotic' and 'exact'.}

\item{threshold}{A \code{character} of length 2. Genes with a percent of
escape lower than the first number will be considered subject to
inactivation. Genes with a higher percent of escape than the second number
will be categorized as escaping inactivation.}

\item{min_support}{A \code{numeric}. Only calculate percentage of escape for
genes that have at least that many samples for which a prediction was made.}

\item{plot}{A \code{logical}. If set to FALSE, do not print the plot and only
return the (invisible) \code{data.table}.}
}
\value{
An invisible \code{data.table} with the following columns
\item{GENE}{Gene symbol}
\item{N_support}{The number of samples for which a prediction was made for that gene}
\item{N_escape}{The number of samples where this gene was found to escape}
\item{pe}{The fraction of escape, N_escape/N_support}
}
\description{
Barplot of the fraction of gene that are 'subject', 'variable' and 'escape'
inactivation from the calls made in XCIR
}
