% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nophase.R
\name{readRNASNPs}
\alias{readRNASNPs}
\title{Read SNPs from RNA-Seq}
\usage{
readRNASNPs(vcf_file)
}
\arguments{
\item{vcf_file}{A \code{character}. The path to a vcf file.}

\item{vcf_file}{The result of a pileup}
}
\description{
Read SNPs from RNA-Seq that have not been phased.
}
\details{
For phased samples, use \code{readXVcf}.
}
