% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readXCI}
\alias{readXCI}
\title{Read a list of known inactivated genes}
\usage{
readXCI(xciGenes = NULL)
}
\arguments{
\item{xciGenes}{A \code{character} or code{NULL}. By defaults, return a
vector of 177 genes. Other available choices include "cotton" and "intersect".
If a file path is given, the genes will be read from
the file.}
}
\value{
A \code{character} vector of gene names.
}
\description{
Read a list of gene symbols of known inactivated genes
to be used as training set in \code{betaBinomXI}.
}
\details{
The default gene list is from Carrel et al. Nature 2005. doi:10.1038/nature03479.
Cotton, is a list of 294 inactivated genes compiled by Cotton et al. Genome
Biology (2013). doi:10.1186/gb-2013-14-11-r122. Intersect is a more
stringent list that takes the intersection of both Carrel and Cotton's lists.
}
\examples{
xcig <- readXCI()
xcig <- readXCI("cotton")

}
\seealso{
\code{betaBinomXI}
}
